+++ upyog/ml/imports.py
@@ -11,3 +11,8 @@ import torchvision.transforms.functional as TF
 import matplotlib as mpl
 import matplotlib.pyplot as plt
 import seaborn as sns
+
+
+IS_COREML_AVAILABLE = is_package_available("coremltools")
+if IS_COREML_AVAILABLE:
+    import coremltools as ct